% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{search}
\usage{
search(yelp_key, category, city)
}
\arguments{
\item{yelp_key}{a string representing the Yelp API key}

\item{category}{a string representing the category of the business}

\item{city}{a string representing a city name}
}
\value{
a dataframe
}
\description{
Produces a table of 50 businesses
}
\examples{
search("your_yelp_key", "cafe", "Vancouver")
}
